feat(referenceGeometry): implement geometry resolution for reference features. Deduplicate code between features.tst and proxy/execute.ts
